# openapi.yaml
openapi: 3.0.2
info:
  title: Marketplace API
  description: Marketplace API Spec
  version: 1.0.0
servers:
  - url: https://{environment}.marketplace.io
    description: Deployment server
    variables:
      environment:
        default: api
        enum:
          - api           # Production server
          - api.staging   # Staging server
          - api.dev       # Development server
  - url: http://localhost:8080
    description: Local development server
...
paths:
  /categories:
    post:
      tags:
        - category
      summary: Register new category
      operationId: registerNewCategory
      requestBody:
        description: New category payload
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCategory'
      responses:
        '201':
          description: New category registered
          headers:
            Location:
              description: Created resource
              schema:
                type: string
                example: "/categories/dee85bcb"
        '409':
          description: Duplicate category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectReference'
        '422':
          description: Invalid payload
    get:
      tags:
        - category
      summary: Retrieve all categories
      operationId: retrieveAllCategories
      responses:
        '200':
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
              example:
                - id: 'dee85bcb'
                  name: 'Coffee & Tea'
                - id: 'dec734c0'
                  name: 'Snack'
  /categories/{categoryId}:
    get:
      tags:
        - category
      summary: Retrieve category details
      operationId: retrieveCategoryDetails
      parameters: 
        - in: path
          name: categoryId
          description: Category ID
          required: true
          schema:
            type: string
            example: "dee85bcb"
      responses:
        '200':
          description: Category details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
...
components:
  schemas:
    NewCategory:
      type: object
      description: New category payload
      properties:
        name:
          type: string
          example: "Snack"
      required:
        - name
    Category:
      type: object
      description: Category details
      properties:
        id:
          type: string
          example: "dee85bcb"
        name:
          type: string
          example: "Snack"
      required:
        - id
        - name
...
    ObjectReference:
      type: object
      description: Object reference (contains ID only)
      properties:
        id:
          type: string
          example: "dee85bcb"
      required:
        - id